feat(terra-classic-integration): swap tax calc
