+++ util/conversion/conversion.ts
@@ -86,3 +86,15 @@ export const calculateDurationString = (durationInMilli: number): string => {
   return 'imminent'
 }
 export const nanoToMilli = (nano: number) => nano / 1_000_000
+
+
+export const aggregateTaxAmounts = (amounts: { denom: string, amount: string }[])=>amounts.reduce((acc, cur) => {
+    const existingAmount = acc.find((a) => a.denom === cur.denom);
+    if (existingAmount) {
+      existingAmount.amount = (parseFloat(existingAmount.amount) + parseFloat(cur.amount)).toString();
+    } else {
+      acc.push(cur);
+    }
+    return acc;
+  }, [] as { denom: string, amount: string }[]);
+
